chore: Get e2e test harness working again for automated fuzzing; remove old tests that need rewriting
