--- test/e2e/000_e2e.test.js
@@ -1,9 +1,0 @@
-const TestHelpers = require('../TestHelpers')
-
-TestHelpers.run(
-  'e2e',
-  {},
-  (t) => {
-    // This test just confirms that the baseline e2e test is working
-  }
-)
